+++ templates/swift/Example/Podfile
@@ -1,0 +1,11 @@
+use_frameworks!
+
+target '${POD_NAME}_Example', :exclusive => true do
+  pod '${POD_NAME}', :path => '../'
+end
+
+target '${POD_NAME}_Tests', :exclusive => true do
+  pod '${POD_NAME}', :path => '../'
+
+  ${INCLUDED_PODS}
+end
